fix: added password reset templates instead of plain text
