+++ src/backend/app/email_templates/password/password_reset.html
@@ -1,0 +1,84 @@
+<!DOCTYPE html>
+    <html lang="en">
+    <head>
+        <meta charset="UTF-8" />
+        <meta name="viewport" content="width=device-width, initial-scale=1.0" />
+        <style>
+            body {
+                font-family: Arial, sans-serif;
+                background-color: #f4f4f4;
+                margin: 0;
+                padding: 0;
+                display: flex;
+                justify-content: center;
+                align-items: center;
+                height: 100vh;
+            }
+            .email-container {
+                max-width: 600px;
+                background-color: #ffffff;
+                border-radius: 10px;
+                box-shadow: 0 3px 6px rgba(0, 0, 0, 0.1);
+                overflow: hidden;
+                margin: 20px;
+            }
+            .header {
+                background-color: #d73f3f;
+                color: #ffffff;
+                text-align: center;
+                padding: 25px;
+                border-radius: 10px 10px 0 0;
+            }
+            .header h1 {
+                margin: 0;
+                font-size: 26px;
+            }
+            .content {
+                padding: 25px;
+            }
+            .content p {
+                font-size: 16px;
+                line-height: 1.5;
+                color: #333333;
+            }
+            .reset-button {
+                display: inline-block;
+                margin-top: 20px;
+                padding: 12px 24px;
+                background-color: #d73f3f;
+                color: #ffffff;
+                text-decoration: none;
+                border-radius: 5px;
+                transition: background-color 0.3s ease;
+            }
+            .reset-button:hover {
+                background-color: #a33030;
+            }
+            .footer {
+                background-color: #f4f4f4;
+                text-align: center;
+                padding: 20px;
+                color: #666666;
+                font-size: 14px;
+                border-radius: 0 0 10px 10px;
+            }
+        </style>
+    </head>
+    <body>
+        <div class="email-container">
+            <div class="header">
+                <h1>Password Reset Request</h1>
+            </div>
+            <div class="content">
+                <p>Hi {{email}},</p>
+                <p>We received a request to reset your password. Please click the button below to reset it:</p>
+                <a href="{{reset_link}}" class="reset-button">Reset Password</a>
+                <p>If you didn’t request this, please ignore this email.</p>
+                <p>Thank you,<br/>The {{project_name}} Team</p>
+            </div>
+            <div class="footer">
+                <p>Thank you for using Drone Tasking Manager</p>
+            </div>
+        </div>
+    </body>
+    </html>
